requirements.txt
mysql-connector-python~=8.0.22
pytest>=6.0.0
requests
SQLAlchemy~=1.3.20